bot/requirements.txt
httpx>=0.26.0
python-dotenv>=1.0.0
aiogram>=3.3.0
aiohttp>=3.9.0
